Backend/requirements.txt
streamlit
altair
scikit-learn
fastapi
uvicorn
pydantic
vegafusion
pandas
matplotlib
requests
google-cloud-bigquery
db-dtypes